+++ contracts/solidity/contracts/stubs/KeepRandomBeaconOperatorGroupSelectionStub.sol
@@ -15,7 +15,7 @@ contract KeepRandomBeaconOperatorGroupSelectionStub is KeepRandomBeaconOperator 
         return groupSelection.seed;
     }
 
-    function setGroupSize(uint256 size) public {
+    function setGroupSize(uint8 size) public {
         groupSize = size;
         groupSelection.groupSize = size;
     }
